requirements.txt
absl-py==1.2.0
aniso8601==9.0.1
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
colour-science==0.4.1
contourpy==1.0.5
cycler==0.11.0
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3
grpcio==1.49.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.4
imageio==2.22.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.18.0
matplotlib==3.6.1
mysql-connector-python==8.0.30
mysqlclient==2.1.1
numpy==1.23.4
oauthlib==3.2.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.2
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.2
six==1.16.0
SQLAlchemy==1.4.41
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tf-estimator-nightly==2.8.0.dev2021122109
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
